add gen model

+++ md-atom/md-atom-user-core/src/main/java/com/atom/core/model/Service.java
@@ -117,6 +117,6 @@ public class Service {
 	/**
 	 * 
 	 */
-	private int suid;
+	private long suid;
 
 }
