+++ md-atom/md-atom-user-service/src/main/java/com/md/atom/user/service/vo/ServiceVO.java
@@ -114,7 +114,7 @@ public class ServiceVO {
         /**
          *
          */
-        private int suid;
+        private long suid;
 
     }
