+++ md-union/md-union-front-client/src/main/java/com/md/union/front/client/dto/ServiceDTO.java
@@ -114,7 +114,7 @@ public class ServiceDTO {
         /**
          *
          */
-        private int suid;
+        private long suid;
 
     }
